+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Fixed
 - wrong info link in data load screen
 - own data options not working #413
+- table: number formatting and string alignment
 
 ## 5.0.1 - 2024-09-16
 ### Fixed